requirements/base.txt
Django==1.6.1
bpython==0.12
django-braces==1.2.2
django-annoying==0.7.9
django-model-utils==1.5.0
logutils==0.3.3
South==0.8.2
django-suit==0.2.5
django_allauth==0.15.0
pytz